Split out getProductCost
